requirements.txt
# Core Dependencies
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
pillow>=10.0.0

# GPS and Location
gpsd-py3>=0.3.0
geopy>=2.3.0

# Database
sqlalchemy>=2.0.0

# Communication
requests>=2.31.0
python-dotenv>=1.0.0

# Jetson Specific (install separately on Jetson)
# pycuda>=2022.1
# tensorrt>=8.5.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
scikit-learn>=1.3.0
seaborn>=0.12.0

# Web GUI
streamlit>=1.28.0
plotly>=5.17.0
